Moved merge_dicts to the config class
Config class should be responsible for merging dicts, and the strategy
used.
