+++ molecule/__init__.py
@@ -18,7 +18,6 @@
 #  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 #  THE SOFTWARE.
 
-import anyconfig
 import pbr.version
 
 try:
@@ -26,7 +25,3 @@ try:
     __version__ = pbr.version_info.release_string()
 except AttributeError:
     __version__ = None
-
-
-class MoleculeSettings(object):
-    MOLECULE_MERGE_STRATEGY = anyconfig.MS_DICTS
