+++ test/unit/test_config.py
@@ -154,3 +154,13 @@ def test_combine_local_config_overrides_default_and_project_config(temp_files):
 
     assert 'local-override' == config_instance['foo']
     assert 'local-override' == config_instance['baz']
+
+
+def test_merge_dicts():
+    # Example taken from python-anyconfig/anyconfig/__init__.py
+    a = {'b': [{'c': 0}, {'c': 2}], 'd': {'e': 'aaa', 'f': 3}}
+    b = {'a': 1, 'b': [{'c': 3}], 'd': {'e': 'bbb'}}
+    expected = {'a': 1, 'b': [{'c': 3}], 'd': {'e': "bbb", 'f': 3}}
+    result = config.merge_dicts(a, b)
+
+    assert expected == result
