+++ test/unit/test_util.py
@@ -147,16 +147,6 @@ def test_sysexit_with_custom_code():
     assert 2 == e.value.code
 
 
-def test_merge_dicts():
-    # Example taken from python-anyconfig/anyconfig/__init__.py
-    a = {'b': [{'c': 0}, {'c': 2}], 'd': {'e': 'aaa', 'f': 3}}
-    b = {'a': 1, 'b': [{'c': 3}], 'd': {'e': 'bbb'}}
-    expected = {'a': 1, 'b': [{'c': 3}], 'd': {'e': "bbb", 'f': 3}}
-    result = util.merge_dicts(a, b)
-
-    assert expected == result
-
-
 def test_get_cookiecutter_template_dir():
     result = util._get_cookiecutter_template_dir('foo')
     parts = pytest.helpers.os_split(result)
